+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   * Add marquee functionality the header of all modals, if the header text doesn't fit it will now scroll instead of pushing all the content down
   * Only allow one FM or Bluetooth stream to run at once
   * Change how errors are displayed on Stream create/edit modal to be more obvious
+  * Convert Zones settings dropdowns to modals, like all other similar settings pages
 * System
   * Fixed a bug where support tunnel addresses would be served when querying for amplipi.local
   * Stream validation for URLs has been made more robust